Move mapValueToObject to utils.object
