+++ lib/core/options/table.js
@@ -67,16 +67,3 @@ export function getMatchingValueOrAdd<T: { [key: string]: mixed }>(
 
 	return matchingValue;
 }
-
-export function mapValueToObject<T: any[]>(option: TableOption<T>, value: T[]): { [key: string]: mixed } {
-	const object = {};
-
-	option.fields.forEach((field, index) => {
-		Reflect.defineProperty(object, field.key, {
-			get: () => value[index],
-			set: v => { value[index] = v; return value[index]; },
-		});
-	});
-
-	return object;
-}
