+++ lib/modules/featurePenaltyBox.js
@@ -112,7 +112,7 @@ export function penalizedDelay(moduleID: string, featureID: string, delay: Delay
 
 function getOrAddFeatures(moduleID, featureID) {
 	const value = Options.table.getMatchingValueOrAdd(module, 'features', { moduleID, featureID });
-	const obj = Options.table.mapValueToObject(module.options.features, value);
+	const obj = mapValueToObject(module.options.features, value);
 	obj.penalty = parseInt(obj.penalty, 10) || 0;
 	return obj;
 }
