+++ lib/modules/notifications.js
@@ -7,7 +7,7 @@ import * as Init from '../core/init';
 import * as Modules from '../core/modules';
 import * as Options from '../core/options';
 import { Permissions, Storage, i18n } from '../environment';
-import { firstValid, hashCode } from '../utils';
+import { firstValid, hashCode, mapValueToObject } from '../utils';
 import * as CommandLine from './commandLine';
 import * as SettingsNavigation from './settingsNavigation';
 
@@ -128,7 +128,7 @@ function getOrAddNotificationType(notification) {
 	if (typeof value[4] !== 'number') {
 		value[4] = parseInt(value[4], 10) || 0;
 	}
-	return Options.table.mapValueToObject(module.options.notificationTypes, value);
+	return mapValueToObject(module.options.notificationTypes, value);
 }
 
 function enableNotificationType(notificationType, enabled) {
