+++ lib/utils/__tests__/object.js
@@ -2,7 +2,10 @@
 
 import test from 'ava';
 
-import { extendDeep } from '../object';
+import {
+	extendDeep,
+	mapValueToObject,
+ } from '../object';
 
 test('extendDeep', t => {
 	const e = extendDeep;
@@ -14,4 +17,20 @@ test('extendDeep', t => {
 	t.deepEqual(e({ foo: null, baz: 3 }, { foo: { bar: 1 } }), { foo: { bar: 1 }, baz: 3 }, 'handles replacing null');
 	t.deepEqual(e({ foo: { bar: 1 }, baz: 3 }, { foo: undefined }), { foo: undefined, baz: 3 }, 'handles replacing with undefined');
 	t.deepEqual(e({ foo: undefined, baz: 3 }, { foo: { bar: 1 } }), { foo: { bar: 1 }, baz: 3 }, 'handles replacing undefined');
+});
+
+test('mapValueToObject get', t => {
+	const fields = [{ name: 'aaa' }, { name: 'b b' }, { name: 'c_c' }, { name: 'D d' }];
+	const value = [0, 1, 2];
+	const obj = mapValueToObject({ fields }, value);
+	t.deepEqual(obj, { aaa: 0, 'b b': 1, c_c: 2, 'D d': undefined });
+});
+
+test('mapValueToObject set', t => {
+	const fields = [{ name: 'a' }, { name: 'b' }, { name: 'c' }];
+	const value = [0, 0, 0];
+	const obj = mapValueToObject({ fields }, value);
+	obj.a = 1;
+	obj.b = 2;
+	t.deepEqual(obj, { a: 1, b: 2, c: 0 });
 });
