+++ lib/utils/index.js
@@ -121,6 +121,7 @@ export {
 
 export {
 	extendDeep,
+	mapValueToObject,
 } from './object';
 
 export {
