+++ lib/utils/object.js
@@ -20,3 +20,23 @@ export function extendDeep(target: { [key: string]: mixed }, source: { [key: str
 	}
 	return target;
 }
+
+type FieldMapping = {
+	name: string,
+};
+
+type FieldMappingCollection = Array<FieldMapping>;
+
+export function mapValueToObject<T: any[]>(option: FieldMappingCollection, value: T[]): { [key: string]: mixed } {
+	const object = {};
+
+	option.fields.forEach((field, index) => {
+		Reflect.defineProperty(object, field.name, {
+			get: () => value[index],
+			set: v => { value[index] = v; return value[index]; },
+		});
+	});
+
+	return object;
+}
+
